Allow for quantity=0 for negative values too.
When a Split has either quantity=0 or value=0, don't validate their signs. While
Decimal does distinguish between +0 and -0, piecash does not preserve that
information, and it doesn't make much sense to assign significance to the sign
of a zero anyway.

A Split with quantity=0 and value<0 can appear when shares are sold with a loss.

A Split with quantity<0 and value=0 represents giving shares away for free.