Remove `constructor` from `@as_declarative`
sqlalchemy 1.3 defined a `constructor` keyword to the `@as_declarative`
decorator, which is no longer supported in 1.4. We previously used this
to override the default constructor with our own.

The first fix is to define the constructor for DeclarativeBase. This works
out-of-the-box with SA 1.4 with the default @as_declarative.

In SA 1.3, the default behavior of @as_declarative is to provide its
own constructor unless `constructor=None` is given). Adding this keyword
breaks SA >=1.4.

To maintain the current behavior across both versions, I'm using an
import test to determine the version of SA. SA < 1.4 placed
`as_declarative` in `sqlalchemy.ext.declarative` while SA >= 1.4 moved
it into `sqlalchemy.orm` (with a copy available in ext.declarative). If
we sucessfully import from `sqlalchemy.orm`, we know we're on SA >= 1.4
and can skip the `constructor` keyword. Otherwise, we explicitly specify
`constructor=None`.

This passes all tests without errors with SA >=1.4 and < 1.4.