Minimal changes requried to support sqlalchemy 1.4
sqlalchemy 1.4 changed the implementation of `as_declarative`. The new implementation fails to handle
the `constructor` keyword the same way as earlier versions. This commit reimplements `as_declarative` so
that the constructor keyword works as expected in SA 1.4 (and newer).

The allowed version range in `setup.py` is also
incremented appropriately.